day20/b: solution
This was an interesting one. I flirted with the idea of reducing the number of moves in part a.
But then I got the correct solution without the optimization and decided to forge ahead.

I ended up running 1 mix-cycle with the slow code just to get enough data so that I could
verify whether I was making progress with the optimized code.

Signed-off-by: Vassilis Vassiliadis <vasiliadis.vasilis@gmail.com>